requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0